Makefile now size-independent and re-runnable.
Previously, 'make fail' or 'make clean' (as applicable) had to be
run before make could be run again, as the object files etc. created
were moved after there creation to a separate directory (.bin/). Now
all files created by Makefile are created in this directory from the
start. This, and general other cleaning up (e.g. of 'make clean'), means
that 'make' can be called multiple times without fail. The 'make fail'
function is now redundant and has been removed.

More significantly, the Makefile has been modified such that it does
not need to increase with the size of the project. Instead of files or
directories being listed individually as they are made, this process
is now automatic. As linker scripts do not support variables, the
'INPUT' section was mooved to the calling recipe in the Makefile.

TODO: Modify README to account for these changes.

+++ Makefile
@@ -1,18 +1,24 @@
+### Define parameters for compilation etc.
+
 CC = cross-compiler/bin/i686-elf-gcc
 CFLAGS = -ffreestanding -Wall -std=c99
 AS = nasm
 ASFLAGS = -f elf
 LDFLAGS = -nostdlib -ffreestanding -lgcc
-NASM := $(shell command -v nasm 2> /dev/null)
-MKISOFS := $(shell command -v mkisofs 2> /dev/null)
 
-C_SOURCES = $(wildcard os/cpu/*.c os/drivers/*.c os/kernel/*.c)
-ASM_SOURCES = $(wildcard os/cpu/*.asm os/drivers/*.asm os/kernel/*.asm)
-HEADERS = $(wildcard os/cpu/*.h os/drivers/*.h os/kernel/*.h)
+### Locate object/program files for creation of kernel.bin (NOT bootloader.bin)
+
+C_SOURCES = $(shell find os/ -type f -name '*.c' -not -path 'os/boot/*' 2>/dev/null)
+ASM_SOURCES = $(shell find os/ -type f -name '*.asm' -not -path 'os/boot/*' 2>/dev/null)
 
 C_OBJ_FILES = ${C_SOURCES:.c=.o}
 ASM_OBJ_FILES = ${ASM_SOURCES:.asm=.o}
 OBJ_FILES := $(ASM_OBJ_FILES) $(C_OBJ_FILES)
+
+### Check required dependecies exist / are installed
+
+NASM := $(shell command -v nasm 2> /dev/null)
+MKISOFS := $(shell command -v mkisofs 2> /dev/null)
 
 ifndef NASM
 $(error "No nasm in PATH, consider installing nasm. Required for assembly.")
@@ -26,76 +32,41 @@ ifndef MKISOFS
 $(error "No mkisofs in path, consider installing mkisofs. Required for cdrom ISO boot.")
 endif
 
+### Compilation instructions
+
 all: bootloader.bin kernel.bin 
-	cat bootloader.bin kernel.bin > OS.bin
+	cat .bin/bootloader.bin .bin/kernel.bin > .bin/OS.bin
 	# 2880 pages long for El-Torito ISO 9660 floppy disk ISO
-	dd if=/dev/zero of=OS.img bs=512 count=2880 
-	dd if=OS.bin of=OS.img seek=0 conv=notrunc
-	mv .iso iso
-	mv OS.img iso/
-	mkisofs -quiet -V 'MyOS' -input-charset iso8859-1 -o OS.iso -b OS.img -hide OS.img iso/
-	mv .bin bin
-	mv *.bin *.o bin
-	mv .run run
+	dd if=/dev/zero of=.bin/OS.img bs=512 count=2880 
+	dd if=.bin/OS.bin of=.bin/OS.img seek=0 conv=notrunc
+	mkisofs -quiet -V 'MyOS' -input-charset iso8859-1 -o OS.iso -b OS.img -hide OS.img .bin/
+	test -f run || mv .run run
 	chmod +x run
 
 bootloader.bin:
-	$(AS) os/boot/bootloader.asm -f bin -o bootloader.bin
+	$(AS) os/boot/bootloader.asm -f bin -o .bin/bootloader.bin
 
-# notdir gives file names only so can be recogniesed by link.ld
-kernel.bin: $(notdir ${OBJ_FILES}) 
-	$(CC) -T link.ld $(LDFLAGS)
-	chmod -x kernel.bin
+kernel.bin: $(OBJ_FILES) 
+	$(eval OBJ_INPUT_FILES = $(shell find .bin/ -type f -name '*.o' -not -path '.bin/kernel_entry.o'))
+	$(CC) $(OBJ_INPUT_FILES) -T link.ld $(LDFLAGS)
+	chmod -x .bin/kernel.bin
 
-kernel_entry.o:
-	$(AS) os/kernel/kernel_entry.asm $(ASFLAGS) -o $@
+$(C_OBJ_FILES): 
+	$(CC) $(CFLAGS) -c $(basename $@).c -o .bin/$(notdir $@)
 
-kernel.o:
-	$(CC) $(CFLAGS) -c os/kernel/kernel.c -o $@
+$(ASM_OBJ_FILES): 
+	$(AS) $(basename $@).asm $(ASFLAGS) -o .bin/$(notdir $@)
 
-screen.o:
-	$(CC) $(CFLAGS) -c os/drivers/screen.c -o $@
+### Clean up instructions
 
-ioports.o:
-	$(AS) os/cpu/ioports.asm $(ASFLAGS) -o $@
-
-idt.o:
-	$(CC) $(CFLAGS) -c os/cpu/idt.c -o $@
-
-pic.o:
-	$(CC) $(CFLAGS) -c os/cpu/pic.c -o $@
-
-pit.o:
-	$(CC) $(CFLAGS) -c os/cpu/pit.c -o $@
-
-interrupts.o:
-	$(CC) $(CFLAGS) -c os/cpu/interrupts.c -o $@
-
-cpu.o:
-	$(CC) $(CFLAGS) -c os/cpu/cpu.c -o $@
-
-keyboard.o:
-	$(CC) $(CFLAGS) -c os/drivers/keyboard.c -o $@
-
-timer.o:
-	$(CC) $(CFLAGS) -c os/drivers/timer.c -o $@
-
-paging.o:
-	$(AS) os/kernel/paging.asm $(ASFLAGS) -o $@
-
-gdt.o:
-	$(CC) $(CFLAGS) -c os/cpu/gdt.c -o $@
+BIN_FILES = $(wildcard .bin/*.o .bin/*.img .bin/*.bin)
 
 .PHONY: clean
 clean:
-	rm *.iso
-	rm iso/*
-	mv iso .iso
-	rm bin/*
-	mv bin .bin
-	chmod -x run
-	mv run .run
+ifneq ("$(wildcard $(BIN_FILES))", "")
+	rm $(BIN_FILES)
+endif
+	test ! -f OS.iso || rm OS.iso
+	test ! -f run || chmod -x run
+	test ! -f run || mv run .run
 
-.PHONY: fail
-fail:
-	rm *.o *.bin
